Fix Shopify order phone validation
Updated the Shopify order creation function to ensure the phone number is in a valid format before sending it to the Shopify API, preventing errors related to invalid phone numbers.
[skip gpt_engineer]
