+++ src/components/checkout/PaymentForm.tsx
@@ -7,9 +7,8 @@ import { useCart } from "@/contexts/CartContext";
 import { useAuth } from "@/contexts/AuthContext";
 import { formatCurrency } from "@/lib/utils";
 import { supabase } from "@/integrations/supabase/client";
-import { trackCheckoutStep, trackCheckoutCompleted } from "@/lib/analytics/ecommerce";
-import { createShopifyOrder } from "@/lib/shopify/createOrder";
-import { formatPhoneForShopify } from "@/lib/utils/phoneFormatter";
+import { trackCheckoutStep } from "@/lib/analytics/ecommerce";
+import { createOrder } from "@/lib/utils/paymentUtils";
 
 interface PaymentFormProps {
   clientSecret: string;
@@ -118,63 +117,13 @@ const PaymentForm = ({ clientSecret, total, shippingCost }: PaymentFormProps) =>
         throw paymentError;
       }
 
-      // Get Shopify variant IDs for the products
-      const { data: shopifyProducts, error: shopifyError } = await supabase
-        .from('shopify_products')
-        .select('shopify_variant_id, product_id')
-        .in('product_id', items.map(item => item.id));
-
-      if (shopifyError) throw shopifyError;
-
-      // Create a map of our product IDs to Shopify variant IDs
-      const variantMap = new Map(
-        shopifyProducts?.map(sp => [sp.product_id, sp.shopify_variant_id]) || []
-      );
-
-      const phone = localStorage.getItem('phone') || '';
-      const formattedPhone = formatPhoneForShopify(phone);
-
-      // Create order in Shopify with the correct variant IDs
-      await createShopifyOrder({
-        orderId,
-        customer: {
-          first_name: localStorage.getItem('firstName') || '',
-          last_name: localStorage.getItem('lastName') || '',
-          email: user.email,
-          phone: formattedPhone,
-        },
-        shippingAddress: {
-          address1: localStorage.getItem('shipping_address') || '',
-          address2: localStorage.getItem('shipping_address2') || undefined,
-          city: localStorage.getItem('shipping_city') || '',
-          province: localStorage.getItem('shipping_state') || '',
-          zip: localStorage.getItem('shipping_zip') || '',
-          phone: formattedPhone,
-        },
-        lineItems: items.map(item => ({
-          variant_id: variantMap.get(item.id) || '',
-          quantity: item.quantity || 1,
-          price: item.from_price
-        }))
-      });
-
-      // Track successful checkout with email
-      trackCheckoutCompleted(
-        orderId,
+      await createOrder({
+        user,
         items,
-        {
-          firstName: localStorage.getItem('firstName'),
-          lastName: localStorage.getItem('lastName'),
-          phone: formattedPhone,
-          address1: localStorage.getItem('shipping_address'),
-          city: localStorage.getItem('shipping_city'),
-          state: localStorage.getItem('shipping_state'),
-          zipCode: localStorage.getItem('shipping_zip'),
-        },
         total,
         shippingCost,
-        user.email
-      );
+        orderId,
+      });
 
       clearCart();
       
@@ -186,7 +135,7 @@ const PaymentForm = ({ clientSecret, total, shippingCost }: PaymentFormProps) =>
       toast({
         variant: "destructive",
         title: "Error",
-        description: "An unexpected error occurred. Please try again.",
+        description: error instanceof Error ? error.message : "An unexpected error occurred. Please try again.",
       });
     } finally {
       setIsProcessing(false);
